HAAR-1361: pr changes
