+++ backend/controllers/linkedAccountValidation.js
@@ -5,22 +5,6 @@ const validateLinkedUserOptionSelection = (userExists) => {
   }
   return errors
 }
-
-const validateUserForSearch = (existingUsername) => {
-  const errors = []
-  if (!existingUsername) {
-    errors.push({ href: '#existingUsername', text: 'Enter the existing username' })
-  }
-  if (errors.length) return errors
-  if (existingUsername.length < 2) {
-    errors.push({ href: '#existingUsername', text: 'Existing Username must be 2 characters or more' })
-  } else if (existingUsername.length > 30) {
-    errors.push({ href: '#existingUsername', text: 'Existing Username must be 30 characters or less' })
-  }
-  if (errors.length) return errors
-  return errors
-}
-
 const validateLinkedAdminUserCreate = (existingUsername, adminUsername) => {
   const errors = []
   if (!existingUsername) {
@@ -117,7 +101,6 @@ const validateLinkedGeneralUserCreate = (existingUsername, generalUsername, defa
 
 module.exports = {
   validateLinkedUserOptionSelection,
-  validateUserForSearch,
   validateLinkedAdminUserCreate,
   validateLinkedLsaUserCreate,
   validateLinkedGeneralUserCreate,
